% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRfunctions.R
\name{augment.model.table}
\alias{augment.model.table}
\title{Augment CMR model table}
\usage{
augment.model.table(all.models.output)
}
\arguments{
\item{all.models.output}{(required, list) the list with all model outputs and the
\code{model.table} element returned by \code{\link[RMark]{mark.wrapper}()}.}
}
\value{
The modified \code{all.models.output} list with new columns in the \code{model.table}.
}
\description{
Adds numeric model rankings based on AICc, and model numbers to
the \code{model.table} element in the list returned by
\code{\link[RMark]{mark.wrapper}()}, as well as removing ~ from model names. Model
numbers correspond to the order of model output elements in the list before
\code{model.table}.
}
\details{
The \code{model.table} element of the list returned by
\code{\link[RMark]{mark.wrapper}()} has one row per model ordered by AICc rank with
internal \code{mark} model number encoded as the row-number of the
dataframe.

This function does four things to the \code{model.table} element of the
input \code{all.models.output}:

\itemize{

\item {creates a \code{model.rank} column giving the numeric rank of each
model as judged by AICc.}

\item{converts the internal \code{mark} model number encoded in the row
names to a \code{model.num} column}

\item{replaces the row names of \code{model.table} with
\code{1:nrow(model.table)}}

\item{Removes the ~ character from the \code{model} column for prettier
printing} }
}
\author{
Sarah Gutowsky, Dave Fifield
}
