% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_deviance.R
\name{bootstrap.deviance}
\alias{bootstrap.deviance}
\title{Calculate a bootstrapped c-hat from a fitted CJS Mark model}
\usage{
bootstrap.deviance(x, reps, tsm = FALSE)
}
\arguments{
\item{x}{(required) a MARK model object as returned by
\code{\link[RMark]{mark}()}.}

\item{reps}{(required) number of repetitions}

\item{tsm}{(optional, default \code{FALSE})  whether to include a time since
marking effect in the Mark models fit to the simulated capture histories.}
}
\value{
Returns a 1x3 matrix with columns for: bootstrapped c.hat, upper (97.5\%), and lower(2.5\%)
confidence limits.
}
\description{
This function takes a fitted mark model and computes bootstrapped
mean (and 95\% CI) c-hat.
}
\details{
Makes use of \code{\link[CMRhelper]{sims}()} to simulate and
analyse a set capture histories based on the model. This process is
repeated \code{reps} times and a bootstrapped c-hat value is calculated.
}
\author{
Sarah Gutowsky, Greg Robertson
}
