% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRhelper-package.R
\docType{package}
\name{CMRhelper-package}
\alias{CMRhelper}
\alias{CMRhelper-package}
\title{CMRhelper: Helper Functions for Capture-Mark-Recapture Analyses}
\description{
A library of lovingly and tastefully crafted helper functions to assist with CMR analyses. Includes functions copied/modified from The Gentle Introduciton to MARK, (http://www.phidot.org/software/mark/docs/book/) and Kéry & Schaub, 2012. Bayesian Population Analysis.
}
\keyword{internal}
