% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRfunctions.R
\name{create.td}
\alias{create.td}
\title{Create trap dependence}
\usage{
create.td(ch, varname = "td", begin.time = 1)
}
\arguments{
\item{ch}{(required, numeric vector) capture history vector (0/1 values only)}

\item{varname}{(required, character string, default \code{td}) prefix for
variable name}

\item{begin.time}{(required, integer, default 1) time for first occasion}
}
\value{
Returns a dataframe with trap-dependent variables named
    \code{varnamet+1}, ... , \code{varnamet+nocc-1} where \code{t} is
    begin.time and \code{nocc} is the number of occasions.
}
\description{
Creates trap dependence covariates for a capture history
}
\details{
Assumes that the time between occasions is 1.
}
\author{
From RMARK chapter of gentle introduction:
 http://www.phidot.org/software/mark/docs/book/pdf/app_3.pdf
}
