% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_deviance.R
\name{extract.model}
\alias{extract.model}
\title{Extract values from a Mark object}
\usage{
extract.model(x)
}
\arguments{
\item{x}{(required) a MARK model object as returned by
\code{\link[RMark]{mark}()}.}
}
\value{
A list with needed values to simulate a capture history. Generally sent to
\code{\link[CMRhelper]{simul.boot}()} as the next step.
}
\description{
This function takes a fitted CJS Mark model, and extracts relevant details
(numbers marked and recaptured in each group(s) and estimated \eqn{\phi} and \emph{p} values)
needed to simulate capture histories.
}
\details{
This function is designed for relatively straight forward models, although it
can handle age effects, as estimates are extracted directly from their PIMs.
Any individual covariates are ignored, however covariates in the design
matrix would be included, as they are reflected in the PIMs.
}
\author{
Greg Robertson
}
