% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_deviance.R
\name{extract.n.marked}
\alias{extract.n.marked}
\title{Extract number of marked individuals}
\usage{
extract.n.marked(x)
}
\arguments{
\item{x}{(required) a MARK model object as returned by
\code{\link[RMark]{mark}()}.}
}
\value{
Returns a 2D matrix of numbers first marked in each group (row) by occasion
(column).
}
\description{
Extracts numbers first marked at each occasion and in each group
from a model.
}
\author{
Greg Robertson
}
