% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{force.cleanup}
\alias{force.cleanup}
\title{Force cleanup of temporary files}
\usage{
force.cleanup(
  folder = "",
  prefix = "mark",
  recursive = FALSE,
  test.only = FALSE
)
}
\arguments{
\item{folder}{(optional, character vector or length 1) folder to delete files from.
If absent, defaults to the project root as determined by \code{\link[here]{here}()}.
If specified, folder is relative to project root. Note this function
doesn't require the "here" package to be loaded, but rather just installed.}

\item{prefix}{(optional, character vector of length 1) character string
that all temporary files start with, default = "mark".}

\item{recursive}{(optional, boolean, default = FALSE) If TRUE, searches
recursively down through folder hierarchy for files to delete.}

\item{test.only}{(optional, boolean, default = FALSE) If TRUE, returns the
character vector of file names to be deleted without removing them.}
}
\description{
Remove all mark .inp, .res, .out, .tmp, and .vcv files from a
folder.
}
\details{
NOTE this is a little more liberal than \code{\link[RMark]{cleanup}()} in \pkg{RMark} since it will match
ANY file starting with "prefix" that ends with one of ".inp", ".out", etc.
}
\author{
Dave Fifield
}
