% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRfunctions.R
\name{pasty}
\alias{pasty}
\title{Create capture history character strings}
\usage{
pasty(x)
}
\arguments{
\item{x}{(required, numeric) matrix of capture histories (0/1) with one row
per group and one column per number of occasions - 1.}
}
\value{
A vector of capture history strings with length equal to \code{nrow(x)}.
}
\description{
Converts the input capture history matrix into a vector of
capture history strings.
}
\author{
From \href{http://www.phidot.org/software/mark/docs/book/pdf/app_3.pdf}{RMARK} chapter of
\href{http://www.phidot.org/software/mark/docs/book}{program MARK: A Gentle Introduction}
}
