% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRfunctions.R
\name{select_model_results}
\alias{select_model_results}
\title{Select CMR model by model number}
\usage{
select_model_results(model_num)
}
\arguments{
\item{mod.list}{(required) an object of class \code{marklist} as returned by
\link[RMark]{mark.wrapper()}.}
}
\description{
Add numeric model rankings based on AICc to a model table
}
\details{

}
\author{
Greg Robertson
}
