% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap_deviance.R
\name{sims}
\alias{sims}
\title{??}
\usage{
sims(x, reps, tsm = FALSE)
}
\arguments{
\item{x}{(required) a MARK model object as returned by
\link[RMark]{mark}().}
}
\description{
??
}
\details{
Any pertinent details....
}
\author{
???
}
