% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRfunctions.R
\name{simul.cjs}
\alias{simul.cjs}
\title{Simulate CJS data for 1 group at a time}
\usage{
simul.cjs(phi, p, marked, tsm = FALSE)
}
\arguments{
\item{phi}{(required, numeric matrix) matrix of survival probabilities
(as a number between 0 and 1) with one row per group, one column per
number of occasions - 1.}

\item{p}{(required, numeric matrix) matrix of recapture probabilities
(as a number between 0 and 1) with one row per group, one column per
number of occasions - 1.}

\item{marked}{(required, integer, length >= 1) matrix of number of animals
marked with one row per group, and one column per number of occasions - 1.}

\item{tsm}{(optional). Boolean (default \code{FALSE}) indicating whether a
time since marking effect should be included.}
}
\value{
Returns a 2D matrix of capture histories with one row per individual and one
column per occasion.
}
\description{
This function simulates capture histories based on known phi and p.
}
\details{
The code was modifield from the function of the same name presented in
Kéry & Schaub(2012), p.178, with the addition of the tsm argument.
}
\author{
Modified from Kéry & Schaub, 2012. Bayesian Population Analysis, p. 178.
Original code available here:
\url{https://www.vogelwarte.ch/modx/assets/files/publications/BPA/Web\%20appendix\%201.txt}
}
