% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap_deviance.R
\name{bootstrap.deviance}
\alias{bootstrap.deviance}
\title{??}
\usage{
bootstrap.deviance(x, reps, tsm = FALSE)
}
\arguments{
\item{x}{(required) a MARK model object as returned by
\link[RMark]{mark}().}

\item{reps}{(required)}

\item{tsm}{(optional)}
}
\value{
Returns bootstrapped c.hat.
}
\description{
??
}
\details{
Any pertinent details....
}
\author{
Greg Robertson?
}
