% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap_deviance.R
\name{bootstrap.deviance}
\alias{bootstrap.deviance}
\title{Calculate a bootstrapped c-hat value from a fitted CJS mark model}
\usage{
bootstrap.deviance(x, reps, tsm = FALSE)
}
\arguments{
\item{x}{(required) a MARK model object as returned by
\link[RMark]{mark}().}

\item{reps}{(required)}

\item{tsm}{(optional)}
}
\value{
Returns a bootstrapped c.hat.
}
\description{
This function takes a fitted mark model and passes it to
   \link[CMRhelper]{extract.model}() to extract relevant information from the
   model and then passes those to \link[CMRhelper]{sims}() to simulate and
   analyse a set capture history based on the model. This process is
   repeated @param reps times and a bootstrapped c-hat value is calculated.
}
\details{
Any pertinent details....
}
\author{
Sarah Gutowsky, Greg Robertson
}
