% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap_deviance.R
\name{extract.model}
\alias{extract.model}
\title{Extract values from a mark object needed to simulate capture histories}
\usage{
extract.model(x)
}
\arguments{
\item{x}{(required) a MARK model object as returned by
\link[RMark]{mark}().}
}
\value{
A list with needed values to simulate a capture history. Generally sent to
 \link[CMRhelper]{simul.boot}() as the next step.
}
\description{
This function takes a fitted CJS mark model, and extracts relevant details
 (numbers marked and recaptured in each group(s) and estimated \phi and *p* values).
}
\details{
This function is designed for relatively straight forward models, although it
 can handle age effects, as estimates are extracted directly from their PIMs.
 Any individual covariates are ignored, however covariates in the design
 matrix would be included, as they are reflected in the PIMs.
}
\author{
Greg Robertson
}
