% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap_deviance.R
\name{Marked.3}
\alias{Marked.3}
\title{Extract marked individuals - old way}
\usage{
Marked.3(data, n.occasions, groups)
}
\arguments{
\item{data}{(required) the data.}

\item{n.occasions}{(required) number of occasions.}

\item{groups}{(required) a vector of group labels.}
}
\value{
Returns a 2D matrix of capture histories with one row per individual and one
  column per occasion.
}
\description{
Extracts numbers first marked at each occasion and in each group.
}
\details{
This function is deprecated and may go away.

   The default param values (which were specific to the dipper data)
   have been removed. Also fixed a bug where "groups" was referred to as
   "Groups" in the function.
}
\author{
?? - from interweb - need to find source
}
