% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap_deviance.R
\name{Marked}
\alias{Marked}
\title{Extract marked individuals}
\usage{
Marked(x)
}
\arguments{
\item{x}{(required) a MARK model object as returned by
\link[RMark]{mark}().}
}
\value{
Returns a 2D matrix of capture histories with one row per individual and one
  column per occasion.
}
\description{
Extracts numbers first marked at each occasion and in each group
   from a model.
}
\author{
modified from XXXX by Greg Robertson - this was taken from the interweb ...
}
