% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRfunctions.R
\name{select_model_by_rank}
\alias{select_model_by_rank}
\title{Select CMR model by rank}
\usage{
select_model_by_rank(mlist, rank = 1)
}
\arguments{
\item{mod.list}{(required) an object of class \code{marklist} as returned by
\link[RMark]{mark.wrapper()}.}
}
\value{
If \code{rank} is a single integer, then the model with that ranking
    is returned. If \code{rank} is a vector of integers, then a list of models
    with those rankings is returned.
}
\description{
Add numeric model rankings based on AICc to a model table
}
\details{
XXX replace with a more omnibus function called select_model that can select
models by rank, model number, name (all allowed to be length n vectors to
select multiple models). Or select by "phi=time", "p=time | groups | covariates",
"phi=time + tsm", or as a formula phi = ~ time + tsm, or all models whose weight
is > x, or all models within x AIC units of the top model (or even any arbitray
model?)
}
\author{
Dave Fifield based on idea from Sarah Gutowsky?
}
