% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMRfunctions.R
\name{select_model_results}
\alias{select_model_results}
\title{Select model results}
\usage{
select_model_results(model_num)
}
\arguments{
\item{model_num}{(required, integer) the model number of interest}
}
\value{
The chosen model.
}
\description{
Select a particular model of interest from a list of models.
}
\details{
Converts a numeric \code{model_num} offset  to a symbolic model
    name and extracts the named model from global variable \code{all.models} using
    global variable \code{model.list}. This requires the following syntax in
    the calling script:

    \code{all.models = mark.wrapper(model.list, data = data.processed,
       ddl = data.ddl, threads = 2)}

    It would be a good idea to replace these global vars with passed params
    to make this more generic.
}
\author{
Who do?
}
