% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_deviance.R
\name{sims}
\alias{sims}
\title{Perform repeated mark analyses on simulated capture histories}
\usage{
sims(x, reps, tsm = FALSE)
}
\arguments{
\item{x}{(required) a MARK model object as returned by
\code{\link[RMark]{mark}()}.}

\item{reps}{(required) number of repetitions}

\item{tsm}{(optional, default \code{FALSE}) whether to include a time since
marking effect in the Mark models fit to the simulated capture histories.}
}
\value{
A list containing mean deviance along with 95\\% confidence limits.
}
\description{
Use RMark to setup and run multiple analyses
}
\details{
Uses \code{\link[CMRhelper]{simul.boot}()} to generate simulated capture histories
from a fitted model. Then uses \code{RMark} to fit models to these simulated
histories.
}
\author{
Sarah Gutowksy, with further options by Greg Robertson
}
