% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap_deviance.R
\name{simul.boot}
\alias{simul.boot}
\title{??}
\usage{
simul.boot(x)
}
\arguments{
\item{x}{(required) a MARK model object as returned by
\link[RMark]{mark}().}
}
\description{
??
}
\details{
Any pertinent details....
}
\author{
???
}
